Merge NIT-Sofware development features

+++ backend-rest-api/src/main/java/com/nitsoft/ecommerce/service/CompanyService.java
@@ -15,7 +15,7 @@ public class CompanyService {
         return companyRepository.findAll();
     }
 
-    public Company findByCompanyId(Integer companyId) {  
+    public Company findByCompanyId(long companyId) {
         return companyRepository.findByCompanyId(companyId);
     }
 }
